Fix integer overflow in int_place_resolver by casting to bigint
Some numeric card_ids (e.g. 1756812133361540744) exceed the 4-byte
integer max (~2.1B). Cast to bigint instead to handle large IDs.

+++ models/silver/int_place_resolver.sql
@@ -66,7 +66,7 @@ from distinct_card_ids ids
 -- Integer match: card_id is numeric and matches places.place_id
 left join places p_int
     on ids.card_id ~ '^\d+$'
-    and (case when ids.card_id ~ '^\d+$' then ids.card_id::integer end) = p_int.place_id
+    and (case when ids.card_id ~ '^\d+$' then ids.card_id::bigint end) = p_int.place_id
 -- Deck SKU match
 left join places p_sku
     on ids.card_id = p_sku.deck_sku
